// pd-controller/webui/src/app/types.ts
export type UiSourceBundle = {
  rustscript: string;
  javascript: string;
  lua: string;
  scheme: string;
};

export const initialSource: UiSourceBundle = {
  rustscript: "use vm;\n",
  javascript: "import * as vm from \"vm\";\n",
  lua: "local vm = require(\"vm\")\n",
  scheme: "(require (prefix-in vm. \"vm\"))\n"
};

export type UiRenderResponse = { source: UiSourceBundle };

export type UiGraphNode = {
  id: string;
  block_id: string;
  values: Record<string, string>;
  position?: { x: number; y: number };
};

export type UiGraphEdge = {
  source: string;
  source_output: string;
  target: string;
  target_input: string;
};

export type UiGraphEdgeWire = UiGraphEdge & {
  sourceHandle?: string;
  targetHandle?: string;
  data?: {
    source_output?: string;
    target_input?: string;
  };
};

export type ProgramSummary = {
  program_id: string;
  name: string;
  latest_version: number;
  versions: number;
  created_unix_ms: number;
  updated_unix_ms: number;
};

export type ProgramVersionSummary = {
  version: number;
  created_unix_ms: number;
  flavor: string;
  flow_synced: boolean;
};

export type ProgramListResponse = { programs: ProgramSummary[] };

export type ProgramDetailResponse = {
  program_id: string;
  name: string;
  latest_version: number;
  created_unix_ms: number;
  updated_unix_ms: number;
  versions: ProgramVersionSummary[];
};

export type ProgramVersionDetail = {
  version: number;
  created_unix_ms: number;
  flavor: string;
  flow_synced: boolean;
  nodes: UiGraphNode[];
  edges: UiGraphEdge[];
  source: UiSourceBundle;
};

export type ProgramVersionResponse = {
  program_id: string;
  name: string;
  detail: ProgramVersionDetail;
};